29.08.2023

+++ src/main/java/telran/accounting/dto/StatsDto.java
@@ -7,5 +7,5 @@ public class StatsDto {
     protected Integer solvedProblems;
     protected Integer checkedSolutions;
     protected Integer formulatedProblems;
-    protected Integer rating;
+    protected Double rating;
 }
