+++ src/main/java/telran/accounting/model/Stats.java
@@ -13,5 +13,5 @@ public class Stats {
     protected Integer solvedProblems;
     protected Integer checkedSolutions;
     protected Integer formulatedProblems;
-    protected Integer rating;
+    protected Double rating;
 }
